new adminController + changed DTO's

+++ src/main/java/skni/kamilG/skin_sensors_api/Sensor/Model/DTO/SensorResponse.java
@@ -1,5 +1,6 @@
 package skni.kamilG.skin_sensors_api.Sensor.Model.DTO;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import skni.kamilG.skin_sensors_api.Sensor.Model.Location;
 import skni.kamilG.skin_sensors_api.Sensor.Model.SensorStatus;
@@ -8,7 +9,7 @@ public record SensorResponse(
     Short id,
     SensorStatus status,
     ZonedDateTime lastUpdate,
-    Short temperature,
+    BigDecimal temperature,
     Integer humidity,
     Integer pressure,
     Location location,
